Added some resilience4J & save manga to pdf

+++ src/main/java/com/om3g4zell/mangacrawler/entities/AbstractPage.java
@@ -5,7 +5,7 @@ import org.immutables.value.Value;
 @Value.Immutable
 public abstract class AbstractPage {
 
-    abstract int number();
+    abstract double number();
 
     abstract String imageUrl();
 }
